+++ common/src/main/java/com/lowdragmc/lowdraglib/test/ui/TestSceneObject.java
@@ -3,7 +3,6 @@ package com.lowdragmc.lowdraglib.test.ui;
 import com.lowdragmc.lowdraglib.gui.editor.annotation.LDLRegisterClient;
 import com.lowdragmc.lowdraglib.gui.editor.ui.sceneeditor.SceneEditorWidget;
 import com.lowdragmc.lowdraglib.gui.editor.ui.sceneeditor.sceneobject.utils.SizeBoxObject;
-import com.lowdragmc.lowdraglib.gui.editor.ui.sceneeditor.sceneobject.utils.TransformAnchorObject;
 import com.lowdragmc.lowdraglib.gui.modular.IUIHolder;
 import com.lowdragmc.lowdraglib.gui.modular.ModularUI;
 import lombok.NoArgsConstructor;
@@ -35,9 +34,9 @@ public class TestSceneObject implements IUITest {
                 playerPos.south()
         ));
 
-        var anchor = new TransformAnchorObject();
-        anchor.transform().position(new Vector3f(playerPos.above().getX(), playerPos.above().getY(), playerPos.above().getZ()));
-        sceneWidget.addSceneObject(anchor);
+//        var anchor = new TransformAnchorObject();
+//        anchor.transform().position(new Vector3f(playerPos.above().getX(), playerPos.above().getY(), playerPos.above().getZ()));
+//        sceneWidget.addSceneObject(anchor);
 
         var sizeBox = new SizeBoxObject();
         sizeBox.transform().position(new Vector3f(playerPos.below().getX(), playerPos.below().getY() + 1, playerPos.below().getZ()));